Move package balloon/hashing to root
Closes #13.
