+++ balloon/balloon.go
@@ -22,9 +22,9 @@ import (
 	"encoding/binary"
 	"fmt"
 
-	"github.com/bbva/qed/balloon/hashing"
 	"github.com/bbva/qed/balloon/history"
 	"github.com/bbva/qed/balloon/hyper"
+	"github.com/bbva/qed/hashing"
 	"github.com/bbva/qed/log"
 )
