+++ balloon/history/tree.go
@@ -24,8 +24,8 @@ import (
 	"encoding/binary"
 	"math"
 
-	"github.com/bbva/qed/balloon/hashing"
 	"github.com/bbva/qed/balloon/history/storage"
+	"github.com/bbva/qed/hashing"
 	"github.com/bbva/qed/log"
 )
