+++ balloon/history/tree_test.go
@@ -22,7 +22,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/bbva/qed/balloon/hashing"
+	"github.com/bbva/qed/hashing"
 	"github.com/bbva/qed/storage/badger"
 	"github.com/bbva/qed/storage/bplus"
 	"github.com/bbva/qed/testutils/rand"
