+++ balloon/history/verifier.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"math"
 
-	"github.com/bbva/qed/balloon/hashing"
+	"github.com/bbva/qed/hashing"
 	"github.com/bbva/qed/log"
 )
