+++ balloon/hyper/tree.go
@@ -24,8 +24,8 @@ package hyper
 import (
 	"math"
 
-	"github.com/bbva/qed/balloon/hashing"
 	"github.com/bbva/qed/balloon/hyper/storage"
+	"github.com/bbva/qed/hashing"
 	"github.com/bbva/qed/log"
 )
