+++ balloon/hyper/verifier_test.go
@@ -21,8 +21,8 @@ import (
 	"encoding/binary"
 	"testing"
 
-	"github.com/bbva/qed/balloon/hashing"
 	"github.com/bbva/qed/balloon/hyper/storage"
+	"github.com/bbva/qed/hashing"
 	"github.com/bbva/qed/storage/cache"
 )
