+++ balloon/verifier.go
@@ -19,7 +19,7 @@ package balloon
 import (
 	"fmt"
 
-	"github.com/bbva/qed/balloon/hashing"
+	"github.com/bbva/qed/hashing"
 )
 
 type Verifiable interface {
