+++ hashing/hasher.go
